nuke isakmpd_addresses['me']
it turned out that including 'me' in isakmpd_addresses is not convenient
because isakmpd_addresses is best shared by all gateways and 'me' is
specific to a host.

you may define isakmpd_addresses in group_vars like

    isakmpd_addresses:
        foo: 192.168.1.1
        bar: 192.168.2.1
        me: {{ ansible_default_ipv4.address }}

however, it is impossible to listen on different address, such as alias.

isakmpd_listen_address can be defined in group_vars and host_vars
without the restriction.